+++ core_gearworks_lib/source/lib/rendering/renderer/imp_r_renderer.cpp
@@ -4,7 +4,7 @@ gearworks::renderer::renderer() :
 	cur_shader_program_id(0),
 	cur_window(),
 	project_matrix(NULL),
-	view_matrix(glm::translate(mat4(1.0f), vec3(0, 0, 0))),
+	view_matrix(glm::rotate(mat4(1.0f), glm::radians(180.0f), vec3(0, 1, 0))),
 	model_matrix(glm::translate(mat4(1.0f), vec3(0, 0, 0))),
 	current_zoom(1.0f)
 {}
@@ -21,13 +21,13 @@ void gearworks::renderer::update_renderer() {
 
 	// Update the projection matrix so it is also resized along with window/viewport resize
 	// Also factor in zoom for zooming capabilities.
-	project_matrix = glm::ortho(
-		(float)-cur_window->get_winwidth() / current_zoom,
-		(float)cur_window->get_winwidth() / current_zoom,
-		(float)-cur_window->get_winheight() / current_zoom,
-		(float)cur_window->get_winheight() / current_zoom
+	project_matrix = glm::perspective(
+		45.0f,
+		cur_window->get_winasprat(),
+		0.1f,
+		5000.0f
 	);
-
+	
 	// Update the model-view-projection matrix
 	mat4 mvp_matrix = project_matrix * view_matrix * model_matrix;