+++ core_gearworks_lib/source/lib/userdef/engine/u_engine.h
@@ -5,23 +5,6 @@
 
 class gearworks::engine {
 private:
-	// Handle to the renderer
-	gearworks::renderer *renderer_handle;
-public:
-	/// <summary>
-	/// Constructor for the engine class.
-	/// To actually create the object, use the create() function.
-	/// </summary>
-	engine();
-
-	/// <summary>
-	/// Creates an engine object with a given renderer.
-	/// Override this with your own code but be sure to set the renderer_handle variable to the given parameter, _renderer.
-	/// If you don't do that step, other code likely will not work.
-	/// </summary>
-	/// <param name="_renderer">The renderer to `bind` this engine object to.</param>
-	void create(gearworks::renderer *_renderer);
-
 	/// <summary>
 	/// All initializing logic here.
 	/// </summary>
@@ -38,6 +21,42 @@ public:
 	/// Delete any heap allocated members or anything else here.
 	/// </summary>
 	void clean();
+public:
+	engine();
+	~engine();
+
+	/// <summary>
+	/// The main renderer used for the program.
+	/// </summary>
+	gearworks::renderer *renderer = nullptr;
+	/// <summary>
+	/// The main input manager used for the program.
+	/// </summary>
+	gearworks::input_manager *input = nullptr;
+
+	/// <summary>
+	/// Path to the vertex shader to use in the program.
+	/// Make sure to set it BEFORE calling initialize()!
+	std::string vertex_shader_path;
+	/// </summary>
+	/// <summary>
+	/// Path to the fragment shader to use in the program.
+	/// Make sure to set it BEFORE calling initialize()!
+	/// </summary>
+	std::string fragment_shader_path;
+
+	/// <summary>
+	/// Initializes the engine process and its members.
+	/// </summary>
+	void behaviour_initialize();
+	/// <summary>
+	/// Updates the engine process and its members.
+	/// </summary>
+	void behaviour_update();
+	/// <summary>
+	/// Terminates the engine process and its members.
+	/// </summary>
+	void behaviour_destroy();
 };
 
 #endif // header guard